Refactor domain package
