+++ internal/application/scanner/domain/entities.go
@@ -1,5 +1,7 @@
 package scanner
 
+// TODO: validation
+// TODO: create custom response objects for domain entities
 type Scanner struct {
 	ID       string
 	Name     string
